requirement.txt
Pillow==7.1.2
aiohttp==3.6.2
pytest-aiohttp==0.3.0
asyncio==3.4.3
aiohttp-jinja2==1.2.0